Move ORM Query Adapter to package
