+++ composer.json
@@ -28,7 +28,8 @@
             "Pagerfanta\\": "lib/Core/",
             "Pagerfanta\\Adapter\\": "src/Adapter/",
             "Pagerfanta\\DoctrineCollections\\": "lib/Adapter/DoctrineCollections/",
-            "Pagerfanta\\DoctrineDBAL\\": "lib/Adapter/DoctrineDBAL/"
+            "Pagerfanta\\DoctrineDBAL\\": "lib/Adapter/DoctrineDBAL/",
+            "Pagerfanta\\DoctrineORM\\": "lib/Adapter/DoctrineORM/"
         }
     },
     "autoload-dev": {
