+++ lib/Adapter/DoctrineORM/composer.json
@@ -1,0 +1,23 @@
+{
+    "name": "pagerfanta/doctrine-orm-adapter",
+    "type": "library",
+    "description": "Pagerfanta adapter for Doctrine ORM",
+    "keywords": ["pagerfanta", "doctrine", "orm"],
+    "license": "MIT",
+    "require": {
+        "php": "^7.2",
+        "doctrine/orm": "^2.5",
+        "pagerfanta/core": "^2.4"
+    },
+    "autoload": {
+        "psr-4": {
+            "Pagerfanta\\DoctrineORM\\": "./"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-2.x": "2.x-dev"
+        }
+    },
+    "minimum-stability": "dev"
+}
